split auth service to main service and utils
