+++ .env.example
@@ -55,3 +55,6 @@ RABBITMQ_DEFAULT_PASS=techcell_server_9901
 REDIS_HOST=localhost
 REDIS_PORT=6379
 REDIS_PASSWORD=techcell_server_redis_password_9901
+
+GOOGLE_CLIENT_ID= <your-client-id>
+GOOGLE_CLIENT_SECRET= <your-client-secret>
