+++ apps/auth/src/auth.module.ts
@@ -3,7 +3,7 @@ import { RabbitMQModule, RabbitMQService } from '@app/common';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { UsersModule } from './users/users.module';
-import { AccessTokenStrategy } from './strategies';
+import { AccessTokenStrategy, GoogleStrategy } from './strategies';
 import { JwtModule } from '@nestjs/jwt';
 import { JwtGuard } from './guards';
 import { MAIL_SERVICE } from '~/constants';
@@ -18,6 +18,6 @@ import { OtpModule } from './otp';
         RabbitMQModule.registerRmq(MAIL_SERVICE, process.env.RABBITMQ_MAIL_QUEUE),
     ],
     controllers: [AuthController],
-    providers: [JwtGuard, RabbitMQService, AuthService, AccessTokenStrategy],
+    providers: [JwtGuard, RabbitMQService, AuthService, AccessTokenStrategy, GoogleStrategy],
 })
 export class AuthModule {}
