+++ apps/auth/src/dtos/check-email-request.dto.ts
