+++ apps/auth/src/guards/google.guard.ts
@@ -1,0 +1,12 @@
+import { Injectable } from '@nestjs/common';
+import { ConfigService } from '@nestjs/config';
+import { AuthGuard } from '@nestjs/passport';
+
+@Injectable()
+export class GoogleOAuthGuard extends AuthGuard('google') {
+    constructor(private configService: ConfigService) {
+        super({
+            accessType: 'offline',
+        });
+    }
+}
