+++ apps/auth/src/guards/index.ts
@@ -1,1 +1,2 @@
 export * from './jwt.guard';
+export * from './google.guard';
