+++ apps/auth/src/strategies/index.ts
@@ -1,2 +1,2 @@
 export * from './accessToken.strategy';
-// export * from './refreshToken.strategy';
+export * from './google.strategy';
