+++ package.json
@@ -44,7 +44,7 @@
     "@nestjs/jwt": "^10.0.3",
     "@nestjs/microservices": "^9.4.2",
     "@nestjs/mongoose": "^9.2.2",
-    "@nestjs/passport": "^9.0.3",
+    "@nestjs/passport": "^10.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/swagger": "^6.3.0",
     "@nestjs/throttler": "^4.1.0",
@@ -66,6 +66,7 @@
     "nodemailer": "^6.9.3",
     "otp-generator": "^4.0.1",
     "passport": "^0.6.0",
+    "passport-google-oauth20": "^2.0.0",
     "passport-jwt": "^4.0.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.2.0"
@@ -81,6 +82,7 @@
     "@types/node": "18.16.12",
     "@types/nodemailer": "^6.4.8",
     "@types/otp-generator": "^4.0.0",
+    "@types/passport-google-oauth20": "^2.0.11",
     "@types/supertest": "^2.0.11",
     "@typescript-eslint/eslint-plugin": "^5.0.0",
     "@typescript-eslint/parser": "^5.0.0",
